+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## 6.0.0 (Unreleased)
 
-TODO
+### Changed
+
+-   `IconClass` prop for most components has been renamed to `icon` for clarification and has bee extended to support a wider variety of icon formats (see TODO).
+-   In `<Hero>` component, `ValueIconClass` prop has been renamed to `valueIcon`.
+-   In `<ChannelValue>` component, `IconProps` has been replaced with separate props for `iconSize` and `iconColor`.
+
 
 ## 5.2.0 (July 28, 2021)